add ci test against PR
