+++ .bumpversion.cfg
@@ -1,7 +1,7 @@
 [bumpversion]
 commit = False
 tag = True
-current_version = 1.0.129
+current_version = 1.0.130
 tag_name = v{current_version}
 message = GitHub Actions Build {current_version}
