+++ .github/workflows/Test_CI.yml
@@ -6,14 +6,10 @@ name: Test CI
 on:
   push:
     branches: [ main ]
-  #pull_request:
-  #  branches: [ main ]
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
-
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python 3.9
@@ -24,8 +20,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest pytest-cov
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        if [ -f tests_requirements.txt ]; then pip install -r tests_requirements.txt; fi
+        if [[ -f requirements.txt ]]; then pip install -r requirements.txt; fi
+        if [[ -f tests_requirements.txt ]]; then pip install -r tests_requirements.txt; fi
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
